refactor(ai-utils): Extract shared agent assembly from handleAgentChat.
Request: Prepare the agent runtime for a headless run-to-completion mode
alongside streaming chat.
Motivation: The agent assembly (tools, model, hooks, stop conditions,
ToolLoopAgent construction) was embedded in the streaming path; a headless
mode needs the same assembly without duplicating it.

Decisions:
- Pure extraction, no behavior change: createToolLoopAgent returns
  { agentInstance, mcpClients, model, timeoutConfig, locale } and contains
  the moved cleanHookEvent/hookMapping/createHookCallbacks helpers.
- createUsageAccumulator moved unchanged to its own module so both paths
  report the identical usage shape.

Changes:
- createToolLoopAgent.js: extracted agent assembly (tools, update-page-state,
  model, hook callbacks, pageContext instructions, stopWhen, ToolLoopAgent).
- createUsageAccumulator.js: moved accumulator.
- handleAgentChat.js: consumes the extracted modules; streaming logic
  unchanged.

Tags: ai-utils, refactor, agents, tool-loop-agent, extraction

$30m
