+++ packages/utils/ai-utils/src/handleAgentChat.js
@@ -15,95 +15,19 @@
 */
 
 import {
-  ToolLoopAgent,
   convertToModelMessages,
   createAgentUIStream,
   createUIMessageStream,
   createUIMessageStreamResponse,
   generateText,
   pruneMessages,
-  stepCountIs,
-  hasToolCall,
   validateUIMessages,
 } from 'ai';
 
 import { serializer } from '@lowdefy/helpers';
 
-import buildAgentTools from './buildAgentTools.js';
-import buildPrepareStep from './buildPrepareStep.js';
-import buildUpdatePageStateTool from './buildUpdatePageStateTool.js';
-
-function createUsageAccumulator() {
-  const usage = {
-    inputTokens: 0,
-    outputTokens: 0,
-    totalTokens: 0,
-    reasoningTokens: 0,
-    cacheReadTokens: 0,
-    cacheWriteTokens: 0,
-  };
-
-  let finishReason = 'stop';
-
-  function add(stepResult) {
-    const stepUsage = stepResult?.usage ?? stepResult;
-    if (stepResult?.finishReason) {
-      finishReason = stepResult.finishReason;
-    }
-    if (!stepUsage) return;
-    usage.inputTokens += stepUsage.inputTokens ?? 0;
-    usage.outputTokens += stepUsage.outputTokens ?? 0;
-    usage.totalTokens += stepUsage.totalTokens ?? 0;
-    usage.reasoningTokens += stepUsage.outputTokenDetails?.reasoningTokens ?? 0;
-    usage.cacheReadTokens += stepUsage.inputTokenDetails?.cacheReadTokens ?? 0;
-    usage.cacheWriteTokens += stepUsage.inputTokenDetails?.cacheWriteTokens ?? 0;
-  }
-
-  return { usage, add, getFinishReason: () => finishReason };
-}
-
-// Strip non-serializable fields from agent-level hook events before sending as payload.
-// messages excluded here — the stream-level onFinish sends UIMessage[] directly.
-function cleanHookEvent(event) {
-  const clean = {};
-  for (const [key, value] of Object.entries(event)) {
-    if (key === 'abortSignal') continue;
-    if (key === 'messages') continue;
-    if (typeof value === 'function') continue;
-    clean[key] = value;
-  }
-  return clean;
-}
-
-// Maps YAML hook names to AI SDK callback names and creates fire-and-forget callbacks.
-// onFinish is intentionally excluded — it is handled at the stream level inside the
-// createUIMessageStream execute function so that hooks are awaited and can return
-// dataParts to be written to the response stream.
-const hookMapping = {
-  onStart: 'experimental_onStart',
-  onStepStart: 'experimental_onStepStart',
-  onToolCallStart: 'experimental_onToolCallStart',
-  onToolCallFinish: 'experimental_onToolCallFinish',
-  onStepFinish: 'onStepFinish',
-};
-
-function createHookCallbacks({ hooks, callEndpoint, locale }) {
-  if (!hooks) return {};
-
-  const callbacks = {};
-  for (const [yamlKey, sdkKey] of Object.entries(hookMapping)) {
-    const endpointIds = hooks[yamlKey];
-    if (!endpointIds || endpointIds.length === 0) continue;
-
-    callbacks[sdkKey] = (event) => {
-      const payload = { ...cleanHookEvent(event), locale };
-      for (const endpointId of endpointIds) {
-        callEndpoint(endpointId, { payload }).catch(() => {});
-      }
-    };
-  }
-  return callbacks;
-}
+import createToolLoopAgent from './createToolLoopAgent.js';
+import createUsageAccumulator from './createUsageAccumulator.js';
 
 // Convert data: URLs in file parts to raw base64 so the AI SDK does not attempt
 // to download them (it only supports http/https).  The mediaType field already
@@ -138,79 +62,10 @@ async function handleAgentChat({ connection, properties, context }) {
   const { agent, messages: rawMessages } = properties;
   const messages = convertDataUrlsToBase64(rawMessages);
 
-  const { tools, mcpClients } = await buildAgentTools({ agent, context });
-
-  const sharedState = context.agentContext?.sharedState;
-  const updatePageStateTool = buildUpdatePageStateTool({ sharedState });
-  if (updatePageStateTool) {
-    tools['update-page-state'] = updatePageStateTool;
-  }
-
-  const model = connection.provider(agent.properties.model);
-
-  const locale = context.i18n?.active;
-  const hookCallbacks = createHookCallbacks({
-    hooks: agent.hooks,
-    callEndpoint: context.callEndpoint,
-    locale,
-  });
-
-  // Prepend page context to instructions when pageContext is enabled
-  let instructions = agent.properties.instructions;
-  if (agent.properties.pageContext && context.agentContext) {
-    const ctx = context.agentContext;
-    const contextLines = ['<context>'];
-    if (ctx.pageId) contextLines.push(`  pageId: ${ctx.pageId}`);
-    if (ctx.userId) contextLines.push(`  userId: ${ctx.userId}`);
-    if (ctx.conversationId) contextLines.push(`  conversationId: ${ctx.conversationId}`);
-    if (ctx.urlQuery && Object.keys(ctx.urlQuery).length > 0) {
-      contextLines.push(`  urlQuery: ${JSON.stringify(ctx.urlQuery)}`);
-    }
-    if (ctx.sharedState && Object.keys(ctx.sharedState).length > 0) {
-      contextLines.push(`  sharedState: ${JSON.stringify(ctx.sharedState)}`);
-    }
-    contextLines.push('</context>');
-    instructions = `${contextLines.join('\n')}\n\n${instructions ?? ''}`;
-  }
-
-  // Build stop conditions
-  const stopConditions = [stepCountIs(agent.properties.maxSteps ?? 10)];
-  const stopOnToolCall = agent.properties.stopOnToolCall;
-  if (stopOnToolCall) {
-    const toolNames = Array.isArray(stopOnToolCall) ? stopOnToolCall : [stopOnToolCall];
-    for (const name of toolNames) {
-      stopConditions.push(hasToolCall(name));
-    }
-  }
-
-  const agentInstance = new ToolLoopAgent({
-    model,
-    instructions,
-    tools,
-    stopWhen: stopConditions.length === 1 ? stopConditions[0] : stopConditions,
-    maxOutputTokens: agent.properties.maxOutputTokens,
-    temperature: agent.properties.temperature,
-    toolChoice: agent.properties.toolChoice ?? 'auto',
-    providerOptions: agent.properties.providerOptions,
-    activeTools: agent.properties.activeTools,
-    topP: agent.properties.topP,
-    topK: agent.properties.topK,
-    frequencyPenalty: agent.properties.frequencyPenalty,
-    presencePenalty: agent.properties.presencePenalty,
-    seed: agent.properties.seed,
-    stopSequences: agent.properties.stopSequences,
-    maxRetries: agent.properties.maxRetries,
-    ...(agent.properties.prepareStep
-      ? { prepareStep: buildPrepareStep(agent.properties.prepareStep) }
-      : {}),
-    ...hookCallbacks,
-    ...(agent.properties.repairToolCall
-      ? {
-          experimental_repairToolCall: async ({ toolCall }) => {
-            return { ...toolCall };
-          },
-        }
-      : {}),
+  const { agentInstance, mcpClients, model, timeoutConfig, locale } = await createToolLoopAgent({
+    connection,
+    agent,
+    context,
   });
 
   const onFinishEndpointIds = agent.hooks?.onFinish;
@@ -218,8 +73,6 @@ async function handleAgentChat({ connection, properties, context }) {
   const hasMcpClients = mcpClients.length > 0;
 
   const pruneConfig = agent.properties.prune;
-  const timeoutConfig =
-    agent.properties.timeout != null ? { timeout: agent.properties.timeout } : {};
 
   const titleEnabled = agent.properties.generateTitle === true;
